+++ R/2-2.manipulate.R
@@ -444,7 +444,7 @@ c_net_save <- function(go, filename = "net", format = "data.frame") {
       dplyr::select(-1) %>%
       write.csv(., paste0(filename, "_edges.csv"), row.names = FALSE)
   } else if (format == "graphml") {
-    if("id"%in%edge.attributes(go))go <- igraph::delete_edge_attr(go, "id")
+    if ("id" %in% edge.attributes(go)) go <- igraph::delete_edge_attr(go, "id")
     if (!grepl("\\.graphml$", filename)) filename <- paste0(filename, ".graphml")
     igraph::write_graph(go, filename, format = "graphml")
   } else {
@@ -467,19 +467,26 @@ c_net_load <- function(filename, format = "data.frame") {
     edges <- read.csv(paste0(filename, "_edges.csv"), stringsAsFactors = FALSE)
     c_net_from_edgelist(edges, vertex_df = nodes) -> go
   } else if (format == "cyjs") {
-    lib_ps("jsonify",library = FALSE)
+    lib_ps("jsonify", library = FALSE)
     if (!grepl("\\.cyjs$", filename)) filename <- paste0(filename, ".cyjs")
-    jsonify::from_json(filename)->G
-    node=cbind_new(G$elements$nodes$data,G$elements$nodes$position)
-    tmp_fac=max(diff(range(node$x)),diff(range(node$y)))
-    node$x=node$x*2/tmp_fac
-    node$y=-node$y*2/tmp_fac
-    node=node[,colnames(node)!="name"]
-    colnames(node)[1]="name"
-    edge=G$elements$edges$data
-    edge=edge[,!colnames(edge)%in%c("from","to")]
-    colnames(edge)[1:3]=c("id","from","to")
-    c_net_from_edgelist(edge,node)->go
+    jsonify::from_json(filename) -> G
+
+    if (!is.data.frame(G$elements$nodes$data)) {
+      names <- lapply(G$elements$nodes$data, names)
+      comm_name <- Reduce(intersect, names)
+      lapply(G$elements$nodes$data, \(i)i[comm_name]) -> G$elements$nodes$data
+      G$elements$nodes$data <- list_to_dataframe(G$elements$nodes$data)
+    }
+
+    node <- cbind_new(G$elements$nodes$data, G$elements$nodes$position)
+    node$y <- -node$y
+    node <- node[, colnames(node) != "name"]
+    colnames(node)[1] <- "name"
+
+    edge <- G$elements$edges$data
+    edge <- edge[, !colnames(edge) %in% c("from", "to")]
+    colnames(edge)[1:3] <- c("id", "from", "to")
+    c_net_from_edgelist(edge, node) -> go
   } else if (format == "graphml") {
     if (!grepl("\\.graphml$", filename)) filename <- paste0(filename, ".graphml")
     igraph::read_graph(filename, format = "graphml") -> go