+++ R/MetaNet-package.R
@@ -9,7 +9,7 @@
 #' @importFrom utils data combn head tail write.csv packageName
 #' @importFrom stats aggregate median var sd setNames runif relevel coef fitted cor coefficients formula lm pt
 #' @importFrom stats time na.omit kmeans p.adjust density approx ks.test smooth.spline
-#' @importFrom pcutils lib_ps mmscale get_cols trans guolv hebing update_param tidai
+#' @importFrom pcutils lib_ps mmscale get_cols trans guolv hebing update_param tidai change_fac_lev
 #' @importFrom graphics legend hist par text symbols
 #' @importFrom tibble rownames_to_column column_to_rownames
 #' @importFrom reshape2 acast melt